Separating out class into new file
